refactored all code into different files
